Move OchardCircuitVersion out of circuit module to avoid feature gating
